# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: IncentEdge API
  version: 1.0.0
  description: |
    # IncentEdge API Documentation

    IncentEdge is an AI-powered incentive discovery, application, and monetization platform - Infrastructure's Bloomberg Terminal for Incentives.

    ## Features
    - **24,458+ verified incentive programs** across federal, state, local, and utility levels
    - **AI-powered matching** using tiered LLM system (DeepSeek + Claude)
    - **Direct Pay eligibility checking** for IRA Section 6417
    - **Multi-format report generation** (PDF, HTML, JSON)
    - **Real-time analytics** and portfolio tracking

    ## Authentication
    All authenticated endpoints require either:
    - Session-based auth (via Supabase Auth cookies)
    - API Key authentication (via `X-API-Key` header)

    ## Rate Limiting
    Rate limits vary by subscription tier:
    - **Free**: 100 requests/hour
    - **Starter**: 1,000 requests/hour
    - **Professional**: 10,000 requests/hour
    - **Team**: 50,000 requests/hour
    - **Enterprise**: Custom limits

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Error Handling
    The API uses standard HTTP response codes:
    - `200` - Success
    - `201` - Created
    - `400` - Bad Request (validation errors)
...
    - `404` - Not Found
...

...
    ```json
    {
...
servers:
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session token (automatic via cookies)

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    IncentiveProgram:
      type: object
      required:
        - id
        - name
        - category
        - program_type
      properties:
        id:
          type: string
          format: uuid
        external_id:
          type: string
          description: External program identifier
          example: IRA-45L
        name:
          type: string
          description: Program name
          example: Section 45L New Energy Efficient Home Credit
        short_name:
          type: string
          nullable: true
          example: 45L Tax Credit
        description:
          type: string
          description: Full program description
        summary:
          type: string
          description: Brief summary
        program_type:
          type: string
          enum: [tax_credit, tax_deduction, grant, rebate, loan, tax_exemption, tax_abatement]
        category:
          type: string
          enum: [federal, state, local, utility]
        subcategory:
          type: string
          nullable: true
          example: IRA
        sector_types:
          type: array
          items:
            type: string
        technology_types:
          type: array
          items:
            type: string
        building_types:
          type: array
          items:
            type: string
        jurisdiction_level:
          type: string
          enum: [federal, state, local, utility]
        state:
          type: string
          nullable: true
          minLength: 2
          maxLength: 2
        counties:
          type: array
          nullable: true
          items:
            type: string
        municipalities:
          type: array
          nullable: true
          items:
            type: string
        incentive_type:
          type: string
        amount_type:
          type: string
          enum: [fixed, percentage, per_unit, per_kw, per_sqft]
        amount_fixed:
          type: number
          nullable: true
        amount_percentage:
          type: number
          nullable: true
        amount_per_unit:
          type: number
          nullable: true
        amount_per_kw:
          type: number
          nullable: true
        amount_per_sqft:
          type: number
          nullable: true
        amount_min:
          type: number
          nullable: true
        amount_max:
          type: number
          nullable: true
        direct_pay_eligible:
          type: boolean
          description: IRA Section 6417 Direct Pay eligible
        transferable:
          type: boolean
        domestic_content_bonus:
          type: boolean
        energy_community_bonus:
          type: boolean
        prevailing_wage_bonus:
          type: boolean
        low_income_bonus:
          type: boolean
        status:
          type: string
          enum: [active, inactive, expired, pending]
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        application_deadline:
          type: string
          format: date
          nullable: true
        administrator:
          type: string
        administering_agency:
          type: string
        source_url:
          type: string
          format: uri
          nullable: true
        application_url:
          type: string
          format: uri
          nullable: true
        application_complexity:
          type: string
          enum: [low, medium, high]
        typical_processing_days:
          type: integer
          nullable: true
        stackable:
          type: boolean
        eligibility_summary:
          type: string
          nullable: true
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
        popularity_score:
          type: integer
          minimum: 0
          maximum: 100
...
    DashboardStats:
      type: object
      properties:
        totalProjects:
          type: integer
          example: 12
        activeProjects:
          type: integer
          example: 8
        completedProjects:
          type: integer
          example: 3
        onHoldProjects:
          type: integer
          example: 1
        totalDevelopmentCost:
          type: number
          example: 1500000000
        totalPotentialIncentives:
          type: number
          example: 450000000
        totalCapturedIncentives:
          type: number
          example: 125000000
        netCost:
          type: number
          example: 1375000000
        subsidyRate:
          type: number
          description: Percentage of TDC covered by incentives
          example: 30.0
        totalApplications:
          type: integer
          example: 45
        pendingApplications:
          type: integer
          example: 15
        approvedApplications:
          type: integer
          example: 25
        successRate:
          type: number
          description: Approval rate percentage
          example: 83.3
        incentivesByCategory:
          type: object
          properties:
            federal:
              type: number
            state:
              type: number
            local:
              type: number
            utility:
              type: number
        projectsByTier:
          type: object
          properties:
            tier_1_efficient:
              type: integer
            tier_2_high_performance:
              type: integer
            tier_3_net_zero:
              type: integer
            tier_3_triple_net_zero:
              type: integer
        capturedThisMonth:
          type: number
        capturedThisQuarter:
          type: number
        capturedThisYear:
          type: number
paths:
...
  /programs/{id}:
    get:
      summary: Get Program Details
      description: Get detailed information about a specific incentive program
      tags:
        - Programs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Program retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/IncentiveProgram'
        '404':
          description: Program not found
...
  /dashboard:
    get:
      summary: Get Dashboard Stats
      description: Get portfolio-wide KPIs and metrics for the organization
      tags:
        - Dashboard
      security:
        - SessionAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: Dashboard stats retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/DashboardStats'
              examples:
                success:
                  value:
                    data:
                      totalProjects: 12
                      activeProjects: 8
                      completedProjects: 3
                      onHoldProjects: 1
                      totalDevelopmentCost: 1500000000
                      totalPotentialIncentives: 450000000
                      totalCapturedIncentives: 125000000
                      netCost: 1375000000
                      subsidyRate: 30.0
                      totalApplications: 45
                      pendingApplications: 15
                      approvedApplications: 25
                      successRate: 83.3
                      incentivesByCategory:
                        federal: 285000000
                        state: 95000000
                        local: 58000000
                        utility: 12000000
                      projectsByTier:
                        tier_1_efficient: 5
                        tier_2_high_performance: 4
                        tier_3_net_zero: 2
                        tier_3_triple_net_zero: 1
                      capturedThisMonth: 8500000
                      capturedThisQuarter: 45000000
                      capturedThisYear: 125000000
        '401':
          description: Unauthorized
...
  /contact:
    post:
      summary: Submit Contact Form
      description: Submit a contact/support request
      tags:
        - Contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - subject
                - message
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                company:
                  type: string
                phone:
                  type: string
                subject:
                  type: string
                message:
                  type: string
                  minLength: 10
      responses:
        '200':
          description: Message sent successfully
        '400':
          description: Validation error
...